Split macros.c into multiple files.
