+++ right/src/caret_config.c
@@ -2,6 +2,7 @@
 #include "arduino_hid/ConsumerAPI.h"
 #include "arduino_hid/SystemAPI.h"
 #include "macros.h"
+#include "macros_status_buffer.h"
 #include "module.h"
 
 caret_configuration_t remappableModes[] = {
