+++ right/src/config_parser/parse_macro.c
@@ -2,6 +2,7 @@
 #include "config_globals.h"
 #include "str_utils.h"
 #include "macros.h"
+#include "macros_status_buffer.h"
 
 parser_error_t parseKeyMacroAction(config_buffer_t *buffer, macro_action_t *macroAction, serialized_macro_action_type_t macroActionType)
 {
