+++ right/src/macro_recorder.c
@@ -1,6 +1,7 @@
 #include "macro_recorder.h"
 #include "led_display.h"
 #include "macros.h"
+#include "macros_status_buffer.h"
 #include "segment_display.h"
 #include "timer.h"
