+++ right/src/macros_commands.h
@@ -1,0 +1,23 @@
+#ifndef __MACROS_COMMANDS_H__
+#define __MACROS_COMMANDS_H__
+
+// Includes:
+
+    #include <stdint.h>
+    #include <stdbool.h>
+    #include "attributes.h"
+    #include "macros_typedefs.h"
+    #include "str_utils.h"
+
+// Macros:
+
+// Typedefs:
+
+// Variables:
+
+// Functions:
+
+    macro_result_t Macros_ProcessCommandAction(void);
+    macro_result_t Macros_ProcessDelayAction();
+
+#endif
