+++ right/src/macros_keyid_parser.c
@@ -1,5 +1,6 @@
 #include "macros.h"
-#include "macro_keyid_parser.h"
+#include "macros_status_buffer.h"
+#include "macros_keyid_parser.h"
 #include "str_utils.h"
 #include <stddef.h>
 #include <stdint.h>
