+++ right/src/macros_set_command.c
@@ -1,10 +1,11 @@
-#include "macro_set_command.h"
+#include "macros_set_command.h"
 #include "config_parser/parse_config.h"
 #include "config_parser/parse_keymap.h"
 #include "key_states.h"
 #include "layer.h"
 #include "ledmap.h"
 #include "macros.h"
+#include "macros_status_buffer.h"
 #include "module.h"
 #include "secondary_role_driver.h"
 #include "timer.h"
@@ -14,7 +15,7 @@
 #include "led_display.h"
 #include "postponer.h"
 #include "macro_recorder.h"
-#include "macro_shortcut_parser.h"
+#include "macros_shortcut_parser.h"
 #include "str_utils.h"
 #include "utils.h"
 #include "layer_switcher.h"
