+++ right/src/macros_set_command.h
@@ -8,7 +8,7 @@
     #include "key_action.h"
     #include "usb_device_config.h"
     #include "key_states.h"
-    #include "macros_core.h"
+    #include "macros_typedefs.h"
     #include "macros_vars.h"
     #include "macros.h"
