+++ right/src/macros_shortcut_parser.h
