+++ right/src/macros_vars.c
@@ -1,14 +1,15 @@
 #include "macros_vars.h"
 #include "postponer.h"
-#include "macro_keyid_parser.h"
+#include "macros_keyid_parser.h"
 #include "utils.h"
 #include "str_utils.h"
 #include "macros.h"
+#include "macros_status_buffer.h"
 #include <stdint.h>
 #include <stdio.h>
 #include "config_parser/config_globals.h"
 #include "debug.h"
-#include "macro_set_command.h"
+#include "macros_set_command.h"
 
 typedef enum {
     Operator_Add,
