+++ right/src/usb_commands/usb_command_apply_config.c
@@ -6,6 +6,7 @@
 #include "keymap.h"
 #include "macro_events.h"
 #include "macros.h"
+#include "macros_status_buffer.h"
 
 void updateUsbBuffer(uint8_t usbStatusCode, uint16_t parserOffset, parser_stage_t parserStage)
 {
